Refactor Debt - Move Service module registraion to Service project, and wrap as IService Collection extension
